# core-lending-loan-servicing-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-loan-servicing
  description: Loan Servicing Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/loan-servicing
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/loan-servicing-cases/{caseId}/disbursements/{disbursementId}:
    get:
      tags:
        - LoanDisbursement
      summary: Get a disbursement by ID
      operationId: getDisbursementById
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: disbursementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoanDisbursementDTO'
    put:
      tags:
        - LoanDisbursement
      summary: Update a disbursement record
      operationId: updateDisbursement
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: disbursementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanDisbursementDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoanDisbursementDTO'
    delete:
      tags:
        - LoanDisbursement
      summary: Delete a disbursement record
      operationId: deleteDisbursement
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: disbursementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/loan-servicing-cases/{caseId}/disbursements:
    get:
      tags:
        - LoanDisbursement
      summary: List/Search disbursements for a specific servicing case
      operationId: findAllDisbursements
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.loanServicingCaseId
          in: query
          description: Exact filter for loan servicing case id
          required: false
          schema:
            type: string
        - name: filters.transactionId
          in: query
          description: Exact filter for transaction id
          required: false
          schema:
            type: string
        - name: filters.disbursementAmount
          in: query
          description: Exact filter for disbursement amount
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[disbursementAmount].from
          in: query
          description: Filter disbursement amount from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[disbursementAmount].to
          in: query
          description: Filter disbursement amount to value
          required: false
          schema:
            type: string
        - name: filters.disbursementDate
          in: query
          description: Exact filter for disbursement date
          required: false
          schema:
            type: string
        - name: filters.isFinalDisbursement
          in: query
          description: Exact filter for is final disbursement
          required: false
          schema:
            type: boolean
        - name: filters.disbursementMethod
          in: query
          description: Exact filter for disbursement method
          required: false
          schema:
            type: string
            enum:
              - INTERNAL
              - EXTERNAL
        - name: filters.disbursementStatus
          in: query
          description: Exact filter for disbursement status
          required: false
          schema:
            type: string
            enum:
              - PENDING
              - PROCESSING
              - COMPLETED
              - FAILED
              - REVERSED
        - name: filters.paymentProviderId
          in: query
          description: Exact filter for payment provider id
          required: false
          schema:
            type: string
        - name: filters.distributorId
          in: query
          description: Exact filter for distributor id
          required: false
          schema:
            type: string
        - name: filters.distributorAgencyId
          in: query
          description: Exact filter for distributor agency id
          required: false
          schema:
            type: string
        - name: filters.distributorAgentId
          in: query
          description: Exact filter for distributor agent id
          required: false
          schema:
            type: string
        - name: filters.externalTransactionReference
          in: query
          description: Exact filter for external transaction reference
          required: false
          schema:
            type: string
        - name: filters.note
          in: query
          description: Exact filter for note
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseLoanDisbursementDTO'
    post:
      tags:
        - LoanDisbursement
      summary: Create a new disbursement under a servicing case
      operationId: createDisbursement
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanDisbursementDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoanDisbursementDTO'
...
components:
  schemas:
...
    LoanDisbursementDTO:
      required:
        - disbursementAmount
        - disbursementDate
        - disbursementMethod
        - disbursementStatus
        - isFinalDisbursement
        - loanServicingCaseId
        - transactionId
      type: object
      properties:
        loanDisbursementId:
          type: string
          format: uuid
          readOnly: true
        loanServicingCaseId:
          type: string
          format: uuid
        transactionId:
          type: string
          format: uuid
        disbursementAmount:
          minimum: 0.01
          exclusiveMinimum: false
          type: number
        disbursementDate:
          type: string
          format: date
        isFinalDisbursement:
          type: boolean
        disbursementMethod:
          type: string
          enum:
            - INTERNAL
            - EXTERNAL
        disbursementStatus:
          type: string
          enum:
            - PENDING
            - PROCESSING
            - COMPLETED
            - FAILED
            - REVERSED
        paymentProviderId:
          type: string
          format: uuid
        distributorId:
          type: string
          format: uuid
        distributorAgencyId:
          type: string
          format: uuid
        distributorAgentId:
          type: string
          format: uuid
        externalTransactionReference:
          maxLength: 255
          minLength: 0
          type: string
        note:
          maxLength: 500
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    PaginationResponseLoanDisbursementDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/LoanDisbursementDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...